requirements.txt
httpx>=0.27
pyyaml>=6.0
apscheduler>=3.10
requests>=2.31
pytest>=8.0
pytest-mock>=3.14
